🏗️ building: hardening dropdown functionality
